requirements.txt
torch
transformers>=4.41.0
datasets
accelerate
peft
trl
bitsandbytes
scikit-learn
numpy
pandas
pyyaml
tqdm
sentencepiece